test: Move `cleanup` logic to global lib
